Define MediaProvider interface
